MNT: move enable_logging from utils.py into utilities.py
Removes the third helper module flagged in review. enable_logging is a
user-facing convenience, so it belongs in the public utilities module
alongside the other analysis helpers rather than in a standalone utils.py.

- rocketpy.utils.enable_logging -> rocketpy.utilities.enable_logging
- delete rocketpy/utils.py; expose rocketpy.utilities via __init__ (imported
  last, since utilities pulls in Environment/Rocket/encoders)
- move the logging unit tests into test_utilities.py; delete test_utils.py
- update the logging user guide and reference pages
